+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Reduce `tick_rate` default value to 10 milliseconds.
+- Change the constraint type of the `Channel` column to `Length` from `Percentage`.
 
 ### Fixed
 - Don't block when resolving channel ids.